pyproject.toml
[tool.poetry]
name = "banking_api"
version = "0.1.0"
description = "This is for practice"
authors = ["MrMika96 <mikamc1996@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
python-decouple = "^3.7"
drf-spectacular = "^0.25.1"
drf-orjson-renderer = "^1.7.1"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"
Pillow = "^9.4.0"
django-tenants = "^3.4.7"
celery = "^5.2.7"
redis = "^4.4.2"
forex-python = "^1.8"
ruff = "^0.1.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
